TM:PE version 1.7.13 - Implemented a permanent fix to solve problems with stuck vehicles/cims caused by third party mods - Added a button to reset stuck vehicles/cims (see mod settings menu) - AI: Improved lane selection algorithm - Bugfix: AI: Lane merging was not working as expected - Bugfix: Pedestrian light states were sometimes not being stored correctly (thanks to Filip for pointing out this problem)

+++ TLM/TLM/Manager/TrafficLightSimulationManager.cs
@@ -165,7 +165,7 @@ namespace TrafficManager.Manager {
 
 		public void OnLevelUnloading() {
 			NodesWithTrafficLightSimulation = new ushort[0];
-			for (ushort nodeId = 0; nodeId < NetManager.MAX_NODE_COUNT; ++nodeId) {
+			for (uint nodeId = 0; nodeId < NetManager.MAX_NODE_COUNT; ++nodeId) {
 				TrafficLightSimulations[nodeId] = null;
 			}
 		}
